core[patch]: move more tests & test utils to core
