+++ langchain-core/src/runnables/tests/runnable_binding.test.ts
@@ -2,7 +2,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { test } from "@jest/globals";
 import { StringOutputParser } from "../../output_parsers/string.js";
-import { FakeChatModel, FakeStreamingLLM } from "../../utils/testing/lib.js";
+import { FakeChatModel, FakeStreamingLLM } from "../../utils/testing/index.js";
 
 test("Bind kwargs to a runnable", async () => {
   const llm = new FakeChatModel({});
